# `#[inline]` integer parsing functions