@tgross35 2024-11-12
🚀 it is a moderately sized routine so let's just make sure this doesn't cause any significant changes to binary size

@bors try @rust-timer queue