@Noratrieb 2024-11-12
Will cachegrind later but I suspect the regressions are caused by doing more codegen because of the inline attr. Seems OK to me.